Added OpenMP functionality for parallel programming

+++ ComplexNumUtils.cpp
@@ -5,7 +5,7 @@ ComplexNum ComplexNumUtils::Conjugate(ComplexNum cn)
 	return ComplexNum(cn.GetReal(), cn.GetImaginary() * (-1));
 }
 
-double ComplexNumUtils::Abs(ComplexNum cn)
+float ComplexNumUtils::Abs(ComplexNum cn)
 {
 	return sqrt(pow(cn.GetReal(), 2) + pow(cn.GetImaginary(), 2));
 }
